2.1 Page 2: Product Show Page
Let's build a page to view a single product! We'll get there by clicking on this link... or the image or... any of this.
We could use the Vue router so that when we click any of these links, there's no full page refresh. We'll save that for another tutorial in part, because I want to show how Vue can be used as just one piece of a traditional app.

Linking to Another Page
This means that we will make these nice traditional links. The Vue component that renders this lives at assets/components/product-list/product-card.vue. Let's start by wrapping the image in an anchor tag. Add an a, a closing a, indent and, for the href, we can use a computed property to keep things clean: :href=productUrl.
Do the same thing down around the product name: <a :href="productUrl">. And to be extra cool, we can use v-text to print the product name.
That part is totally unnecessary, it's just more hipster... because now the anchor can be self-closing. Oooo.
Finally, for the "View Product" button, we should really change this to a normal anchor tag with an href. But to be extra complicated, let's activate it with a click listener: @click="" and call a method we'll add soon: goToProduct().
To be clear: I would not do this in a real app - it's weird to make normal navigational links use fancy JavaScript... and it's not great for accessibility. I'm just trying to make our life harder for the tutorial. Yay learning!
Head over to the browser and open your debugging tools. Thanks to the dev server, we don't even need to refresh: it's already panicking because the "property or method, productUrl is not defined on the instance".