2.2 Page 2: Product Show Page
Let's add that first. In the component, we already have a computed section. Add another: productUrl. For the return, use the fancy "ticks" so we can say /products/ then ${} with product.id inside. As you'll see in a minute, I've already created a route and controller for this URL.
And, yes: I am hardcoding that URL. You can generate URLs dynamically by using FOSJsRoutingBundle, but I'm fine with hardcoding them: it makes my life simpler. We also talked about this in the first tutorial.
To finish things, we need a method for the @click listener: goToProduct(). At the bottom, add a methods section with goToProduct(). Inside, say window.location = and then this.productUrl to re-use our computed property.
Super nice! Let's try it! When I move back to the browser, it automatically reloaded for me. Click any of the spots and... big error page!

The Product Page
Before I started recording the tutorial, I created a route and controller for this URL. Open src/Controller/ProductController.php. This controller renders every page we've seen so far: the homepage - which lists all the products:

44 lines  src/Controller/ProductController.php
class ProductController extends AbstractController
{
    /**
     * @Route("/", name="app_homepage")
     */
    public function index(CategoryRepository $categoryRepository): Response
    {
        return ->render('product/index.html.twig', [
            'categories' => $categoryRepository->findAll(),
        ]);
    }
}
the category page:

 44 lines  src/Controller/ProductController.php
class ProductController extends AbstractController
{
    /**
     * @Route("/category/{id}", name="app_category")
     */
    public function showCategory(Category $category, IriConverterInterface $iriConverter, CategoryRepository $categoryRepository): Response
    {
        return $this->render('product/index.html.twig', [
            'currentCategoryId' => $iriConverter->getIriFromItem($category),
            'categories' => $categoryRepository->findAll(),
        ]);
    }
}
and - now - the individual product page. This is what's currently being executed.

 44 lines  src/Controller/ProductController.php
class ProductController extends AbstractController
{
    /**
     * @Route("/product/{id}", name="app_product")
     */
    public function showProduct(Product $product): Response
    {
        return $this->render('product/index.html.twig');
    }
}
Notice that all of these render the same template, which is templates/product/index.html.twig.
But, hold on, let's back up. We can make this "product show" page look and work however we want. For example, we could decide that this is going to be a normal, Twig-rendered HTML page: maybe we only needed Vue for the fancy product list. So, we could render a new product/show.html.twig and... be done.
Or, if the page is going to be fancy and Vue would be useful, we could render a totally different Webpack entry that renders a different, new Vue application.
Or... again, we could render the same Vue application that we've been using, but make that smarter by detecting that we're on the product show page and rendering in a different way.
We're still doing full page refreshes, but this is the idea behind a single page application, or even a mini single page application that powers just one section of your site. You have one piece of code that renders different components based on the URL. This is what we're going to do.

The current Vue Setup
As a reminder, here is how the whole thing works currently. In index.html.twig, we're including the products entry, which is the assets/products.js file.

 27 lines  templates/product/index.html.twig
{% block stylesheets %}
    {{ encore_entry_link_tags('products') }}
{% endblock %}
{% block javascripts %}
    {{ encore_entry_script_tags('products') }}
{% endblock %}
This grabs the assets/pages/products.vue component - that's right here - and renders it.

 62 lines  assets/pages/products.vue
<template>
    <div class="container-fluid">
        <div class="row">
            <aside :class="asideClass">
                <sidebar
                    :collapsed="sidebarCollapsed"
                    :current-category-id="currentCategoryId"
                    :categories="categories"
                    @toggle-collapsed="toggleSidebarCollapsed"
                />
            </aside>
            <div :class="contentClass">
                <catalog
                    :current-category-id="currentCategoryId"
                    :categories="categories"
                />
            </div>
        </div>
    </div>
</template>
We're now going to make this component smart enough to either render the product list page or the product show page. Normally that kind of toggling is done with a Vue router, but we're going to do it by passing info from the server, which in some ways, will be trickier.
Anyways, the new product show controller is already set up to use that same index.html.twig template that renders the products entrypoint and ultimately renders the products.vue component. But to get this to work, the template needs a categories variable so it can set a global categories JavaScript variable that we're using.