2.3 Page 2: Product Show Page
To pass that, we'll use the same code as the other controllers. Copy the categories variable from up here. Then, we'll need to autowire the CategoryRepository: CategoryRepository $categoryRepository, add a 2nd argument to render(), and paste!
Now when we refresh... it's alive! Well... it's still executing the exact same Vue app as before.... which renders the product list page... but it's a start.

So next: let's pass some info to Vue to say:

  Hey! We are on the product show page!
We'll then use that info inside products.vue to render something different.