3.1. Current Product Id
So here's the plan: make our top-level Vue component - products.vue - able to render either the product list page or the product show page. Why are we making one component able to render two different pages? One reason is that this is basically how Vue router works: one top-level file leverages the Router to render one of many different components based on the URL. And, in this situation, I also want both pages to have the same sidebar. So we'll be re-using the layout from products.vue.

 62 lines  assets/pages/products.vue
<template>
    <div class="container-fluid">
        <div class="row">
            <aside :class="asideClass">
                <sidebar
                    :collapsed="sidebarCollapsed"
                    :current-category-id="currentCategoryId"
                    :categories="categories"
                    @toggle-collapsed="toggleSidebarCollapsed"
                />
            </aside>
            <div :class="contentClass">
                <catalog
                    :current-category-id="currentCategoryId"
                    :categories="categories"
                />
            </div>
        </div>
    </div>
</template>
Passing a currentProductId Global Variable
So... how can this component determine what the current page is? We could parse the URL... or we could pass that info into Vue.
Look at the Twig template. When we're on a specific category page, for example, if I click "Office Supplies", we set a currentCategoryId global variable, which our Vue component reads to load that category. Now we're going to do the same thing again: add a currentProductId if we're on a product page.
In the controller... in the showCategory() action, the currentCategoryId is not the database id, it's the category IRI - like /api/category/5. We generate that by autowiring the IriConverterInterface service.
Down in showProduct(), do the same thing: add IriConverterInterface $iriConverter and then pass a new variable: currentProductId set to $iriConverter->getIriFromItem() and pass it $product, which is the entity object that Symfony automatically queried for.
By the way, we could pass the entire Product object into the template and then serialize it to JSON - similar to what we did with the categories. That would help us avoid an AJAX call for the product data and let Vue render a bit faster. I'm going to avoid that here... mostly to make our life a bit harder. Again, yay learning!
In the template, copy the currentCategoryId code... then change everything to currentProductId. So if that variable is defined, set a new global JavaScript variable and escape it for JavaScript, just in case.
Cool! So in theory, at the browser, if we click into a product, we should have that global variable. I'll try it in the console: window.currentProductId. Nice!