3.2. Current Product Id
Reading the currentProductId in Vue
So how could we read this in a Vue component? Well... it's a global variable... so we could put window.currentProductId anywhere. But in the last tutorial, we started centralizing these global variables into a services/page-context.js file. As a reminder, services/ is a directory that - for the most part - holds files that make AJAX calls. But in the case of page-context, instead of reading data from AJAX calls, it reads global variables.

Add a new method here, export function getCurrentProductId() and make it return window.currentProductId. We can even impress of friends by adding some documentation.
Beautiful!
To make sure this all works, let's render this value in our Vue component. To do that, we need to make it available in the template... because we can't just call random functions from up here. The easiest - and nicest - way to do that is via a computed prop. In the component, add a new one called currentProductId(). Inside, return getCurrentProductId() and hit tab to auto-complete that.
When it auto-completed, PhpStorm automatically added the new import for me... though I don't love that it put this on multiple lines... I'll fix that.
Anyways, now that we have a currentProductId computed prop, we can use it up in the template right before the catalog component, which lists all the products: {{ currentProductId }}.
I love it! Go check the browser. We don't even need to reload! It's already there.
I can feel the power! Now, how can we render something different in the component based on this value? There are a few options, including the tried and true v-if. But instead, we're going to use something fancier called a dynamic component. That's next.