4.1. The Dynamic Component
Our top level component - products.vue- can now read the currentProductId, which will either be set if we're on a product page or will be null if we're on the page that lists products, which is known as the "catalog" in our code.

 66 lines  assets/pages/products.vue
<template>
    <div class="container-fluid">
        <div class="row">
            <aside :class="asideClass">
                <sidebar
                    :collapsed="sidebarCollapsed"
                    :current-category-id="currentCategoryId"
                    :categories="categories"
                    @toggle-collapsed="toggleSidebarCollapsed"
                />
            </aside>
            <div :class="contentClass">
                Product: {{ currentProductId }}
                <catalog
                    :current-category-id="currentCategoryId"
                    :categories="categories"
                />
            </div>
        </div>
    </div>
</template>
So here's the plan: I want both the product show and the catalog pages to use the same sidebar - this list of categories. But each will render something different for the central content part. The catalog will render the catalog component and the product show page will render something different... like a new component that we'll create now.

Bootstrapping the Product Show Component
Inside assets/components/, create a new file called product-show.vue. Inside, add the template - with just a div and some text to start. Next, add the script tag with export default an object with just name set to ProductShow
Before we worry about rendering catalog or product-show, let's first see if we can render both. Start with import ProductShow from '@/components/product-show'
Then, under components, make that available with ProductShow. Finally, in the template, say <product-show/>
Easy enough! When we check the browser, it's already there. Awesome!