4.2. The Dynamic Component
Conditionally Rendering Components
But of course, we don't want to render both components, we want to render either product-show or catalog based on the currentProductId. How can we do that?
Well, one easy option is to use v-if on each component, like v-if="currentProductId" so that it only renders if currentProductId is set. We would do the opposite for catalog.
That is a fine option, really, a great option! But, since we already know how to use v-if, I want to show you another way to do this. It's really the same behind the scenes, but this other way is even nicer if you need to toggle between more than two components. It's called a dynamic component. Oooo.
To use it, we first need to calculate which component should be rendered. Let's add a computed property to do this - call it currentComponent(). Inside, use the ternary syntax: if currentProductId does not equal null, then we want to render ProductShow. Yep, we're referencing the component variable that we imported. Else, we want to render Catalog.
How can we use this in the template to render whatever it returns? By leveraging a special component called... well actually it's called literally... <component>! Clear out the product stuff and change catalog to component. This special tag can render any component. You tell it what to render via a special is prop. Say :is="currentComponent".
That's it! Test drive time! Back at the browser, let's click "All Products" and... yea! That works! It renders the catalog component. Clicking on any category renders the same... but clicking on a product renders product-show.