4.3. The Dynamic Component
Passing Props to Dynamic Component
So far, this component... isn't very interesting and we're not even passing any props to it. But obviously, it will need to know the currentProductId so that we can eventually make an AJAX call for the full product data. Let's add that prop: props, call it productId, and this will be type: String because it's an IRI string. Also add required: true.
Over in products.vue, hmm. It's a little weird: this component will either render the catalog component or the product-show... and those both need different props. These two props are for Catalog. For now, I guess let's just pass all the possible props needed. Add :product-id="currentProductId".
Oh, and back in product-show.vue, to see if this is working, change the text in the template to render the productId prop.
Over at our browser... yea! It already works!
But... it is weird that - no matter which component we're rendering - we're always passing extra props: the first two are needed only for catalog and the last only for product-show. Is that ok? Ah... not really.
Other than being messy, by default, extra props show up as attributes. We can see this in the DOM: this div has an extra category attribute set to a bunch of weird [object] things. I... don't love that.