4.4. The Dynamic Component
Binding Dynamic Props
No problem. Create another computed property. This time call it currentProps(). Like currentComponent, this will return only the props needed for the current situation. Use the ternary syntax again: if this.currentComponent equals ProductShow, we'll return one set of props, else, we'll return a different set. The first situation needs productId set to this.currentProductId. And then second needs... I think currentCategoryId. Let me check... Yep! It needs currentCategoryId and categories. Oh, but while we use these kebab-case attributes in the template, we'll use camel case down in the computed prop: Vue takes care of normalizing all of that.
So: currentCategoryId set to this.currentCategoryId and categories set to this.categories.
Perfecto! But now... how do we use this? Usually we pass props to a component using v-bind - like v-bind:categories... or just :categories, for short. That allows us to pass in one prop. But this time, we want to bind a bunch of props all at once. How can we do that?
The answer is with v-bind set to an object. So v-bind="" - but no :prop-name - then currentProps.
Let's try it! I'll move over and refresh manually just to be safe. And... it works! No errors, no extra weird attributes and every page seems to be not exploding.
Ok: this is a good situation! We have two separate pages and a blank slate waiting for us to fill in the product details... once we get that data via AJAX. That's next.