6.1. Making the Title Component Less Smart
In the last tutorial, we created a title component, which we're reusing so that we can have a consistent look and styling.

 43 lines  assets/components/title.vue
<template>
    <div :class="$style.component">
        <h1>
            {{ categoryName }}
        </h1>
    </div>
</template>
<script>
export default {
    name: 'Title',
};
</script>
<style lang="scss" module>
.component {
    h1 {
        font-size: 1.7rem;
    }
}
</style>
Cool! In product-show.vue, let's use that! Import TitleComponent from @/components/title, add this into the components key, then up here, instead of our manual <h1>, say <title-component />.
I'm purposely not passing a prop to this yet... and yes, I know it looks weird - like "how will it know what title to render?". Ya see... we have a problem.
Back at the browser, we're on a product page and it... seems to... kind of work. It says "All Products". But more importantly, there's an error!

  Missing required prop categories.

We're apparently supposed to pass a categories prop to title... which is weird, because I'm not sure what categories have to do with printing a title. We better go check out that component. Ah... this component is way too smart: it expects us to pass it the array of all categories and the currentCategoryId. And then it does the logic to figure out if we're on a category page and either prints that category name or "All Products". There's no way for us to make it render anything else.
 43 lines  assets/components/title.vue
<template>
    <div :class="$style.component">
        <h1>
            {{ categoryName }}
        </h1>
    </div>
</template>
<script>
export default {
    name: 'Title',
    props: {
        currentCategoryId: {
            type: String,
            default: null,
        },
        categories: {
            type: Array,
            required: true,
        },
    },
    computed: {
        categoryName() {
            if (this.currentCategoryId === null) {
                return 'All Products';
            }
            const category = this.categories.find((cat) => (cat['@id'] === this.currentCategoryId));
            return category ? category.name : '';
        },
    },
};
What we need to do is convert title into a dumb component that does nothing more than receives props and uses them. This is really a mistake that I made in the last tutorial. We've talked a few times about having dumb components that mostly just render markup and then smart components that do calculations & load data, but don't render much markup. This is not an absolute rule... and I don't always follow it - but it's a nice guide to keep things organized and reusable.