6.2. Making the Title Component Less Smart
Converting to a Dumb Component

Ok: let's make this component less smart! Under props, we only need one: call it text. It will be a String and also required. Then, in the template, instead of categoryName, just render text!
What about all the logic inside the categoryName computed prop? Copy this and delete the entire computed section. Now open assets/components/catalog.vue. This is the one place that currently renders the title component and it is what should be responsible for determining its title text. Down in the component, this doesn't have a computed section yet, so add one after data - computed - and paste categoryName. Both this.currentCategoryId and this.categories are available on this component, so this "should" just work.
Back up top, we can shorten <title-component> significantly: we only need to pass text set to categoryName.
Let's make sure this still works. Back at the browser, click "All Products". The title looks good! Try "Office Supplies" and... perfect!
Thanks to this, in product-show, we're free to pass whatever we want to the title, like :text="product.name".
I love that. And... it even works.
Next: let's bring this page completely to life with a full template and a nice, standalone color selector that we'll soon use to selected a product color before adding it to the cart.